requirements.txt
python-telegram-bot==13.12
requests~=2.28.0
psycopg~=3.0.15